[TAN-515] Move granular permissions into core - tests
