--- back/engines/commercial/granular_permissions/spec/policies/comment_reaction_policy_spec.rb
@@ -1,32 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-describe CommentReactionPolicy do
-  subject(:policy) { described_class.new(user, reaction) }
-
-  let(:scope) { CommentReactionPolicy::Scope.new(user, Reaction) }
-  let(:project) { create(:continuous_project, phase_attrs: { with_permissions: true }) }
-  let(:comment) { create(:comment, post: create(:idea, project: project)) }
-
-  context 'for a mortal user who owns the reaction on a project where commenting is only allowed by admins' do
-    let!(:reaction) { create(:reaction, reactable: comment) }
-    let(:user) { reaction.user }
-
-    before do
-      project.phases.first.permissions
-        .find_by(action: 'commenting_idea')
-        .update!(permitted_by: 'admins_moderators')
-    end
-
-    it { is_expected.to permit(:show) }
-    it { expect { policy.create? }.to raise_error(Pundit::NotAuthorizedError) }
-    it { expect { policy.up? }.to raise_error(Pundit::NotAuthorizedError) }
-    it { is_expected.not_to permit(:down) }
-    it { expect { policy.destroy? }.to raise_error(Pundit::NotAuthorizedError) }
-
-    it 'indexes the reaction' do
-      expect(scope.resolve.size).to eq 1
-    end
-  end
-end
