--- back/engines/commercial/granular_permissions/spec/policies/idea_comment_policy_spec.rb
@@ -1,30 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-describe IdeaCommentPolicy do
-  subject { described_class.new(user, comment) }
-
-  let(:scope) { IdeaCommentPolicy::Scope.new(user, idea.comments) }
-
-  context 'for a mortal user who owns the comment in a project where commenting is not permitted' do
-    let!(:user) { create(:user) }
-    let!(:idea) { create(:idea, project: project) }
-    let!(:comment) { create(:comment, post: idea, author: user) }
-    let!(:project) do
-      create(:continuous_budgeting_project, phase_attrs: { with_permissions: true }).tap do |project|
-        project.phases.first.permissions.find_by(action: 'commenting_idea')
-          .update!(permitted_by: 'admins_moderators')
-      end
-    end
-
-    it { is_expected.to     permit(:show) }
-    it { is_expected.not_to permit(:create) }
-    it { is_expected.not_to permit(:update) }
-    it { is_expected.not_to permit(:destroy) }
-
-    it 'indexes the comment' do
-      expect(scope.resolve.size).to eq 1
-    end
-  end
-end
