--- back/engines/commercial/granular_permissions/spec/policies/idea_policy_spec.rb
@@ -1,71 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-describe IdeaPolicy do
-  subject(:policy) { described_class.new(user, @idea) }
-
-  before do
-    IdeaStatus.create_defaults
-    phase = project.phases.first
-    @idea = create(
-      :idea,
-      project: project,
-      author: author,
-      creation_phase: phase.native_survey? ? phase : nil
-    )
-  end
-
-  let(:scope) { IdeaPolicy::Scope.new(user, project.ideas) }
-  let(:author) { create(:user) }
-  let(:permitted_by) { 'admins_moderators' }
-  let(:participation_method) { 'ideation' }
-  let(:posting_enabled) { true }
-  let(:project) do
-    create(:continuous_project, phase_attrs: { with_permissions: true, posting_enabled: posting_enabled, participation_method: participation_method }).tap do |project|
-      project.phases.first.permissions.find_by(action: 'posting_idea')
-        .update!(permitted_by: permitted_by)
-    end
-  end
-
-  context "for a visitor with posting permissions granted to 'everyone'" do
-    let(:user) { nil }
-    let(:permitted_by) { 'everyone' }
-
-    describe 'in a participation method where everyone can post' do
-      let(:participation_method) { 'native_survey' }
-
-      it { is_expected.not_to permit(:show) }
-      it { is_expected.to permit(:create) }
-      it { is_expected.not_to permit(:update) }
-      it { is_expected.not_to permit(:destroy) }
-    end
-
-    describe 'in a participation method where sign-in is required to post' do
-      let(:participation_method) { 'ideation' }
-
-      it { is_expected.to permit(:show) }
-      it { is_expected.not_to permit(:create) }
-      it { is_expected.not_to permit(:update) }
-      it { is_expected.not_to permit(:destroy) }
-
-      it 'indexes the idea' do
-        expect(scope.resolve.size).to eq 1
-      end
-    end
-  end
-
-  context 'for the author of an idea in a project where posting is not permitted' do
-    let(:user) { author }
-    let(:posting_enabled) { false }
-
-    it { is_expected.to permit(:show) }
-    it { expect { policy.create? }.to raise_error(Pundit::NotAuthorizedError) }
-    it { is_expected.to permit(:update) }
-    it { is_expected.to permit(:destroy) }
-
-    it 'indexes the idea' do
-      expect(scope.resolve.size).to eq 1
-    end
-  end
-end
