+++ back/spec/policies/comment_reaction_policy_spec.rb
@@ -8,7 +8,6 @@ describe CommentReactionPolicy do
   let(:scope) { CommentReactionPolicy::Scope.new(user, Reaction) }
   let(:project) { create(:continuous_project) }
   let(:idea) { create(:idea, project: project, phases: project.phases) }
-
   let(:comment) { create(:comment, post: idea) }
 
   context 'for a visitor' do
@@ -96,13 +95,31 @@ describe CommentReactionPolicy do
   # end
 
   context 'for a mortal user who owns the reaction on a project where commenting is disabled' do
-    let(:project) do
-      project = create(:continuous_project)
-      project.phases.first.update!(commenting_enabled: false)
-      project
-    end
+    let(:project) { create(:continuous_project, phase_attrs: { commenting_enabled: false }) }
     let!(:reaction) { create(:reaction, reactable: comment) }
     let(:user) { reaction.user }
+
+    it { is_expected.to permit(:show) }
+    it { expect { policy.create? }.to raise_error(Pundit::NotAuthorizedError) }
+    it { expect { policy.up? }.to raise_error(Pundit::NotAuthorizedError) }
+    it { is_expected.not_to permit(:down) }
+    it { expect { policy.destroy? }.to raise_error(Pundit::NotAuthorizedError) }
+
+    it 'indexes the reaction' do
+      expect(scope.resolve.size).to eq 1
+    end
+  end
+
+  context 'for a mortal user who owns the reaction on a project where commenting is only allowed by admins' do
+    let(:project) { create(:continuous_project, phase_attrs: { with_permissions: true }) }
+    let!(:reaction) { create(:reaction, reactable: comment) }
+    let(:user) { reaction.user }
+
+    before do
+      project.phases.first.permissions
+        .find_by(action: 'commenting_idea')
+        .update!(permitted_by: 'admins_moderators')
+    end
 
     it { is_expected.to permit(:show) }
     it { expect { policy.create? }.to raise_error(Pundit::NotAuthorizedError) }
